# improve bitmap vector offset to report contiguous groups